Move code from Enigma class to Shift class; most corresponding tests
