+++ test/enigma_test.rb
@@ -16,42 +16,6 @@ class EnigmaTest < MiniTest::Test
     assert_equal 27, enigma.characters.size
   end
 
-  def test_it_can_generate_key_string
-    enigma = Enigma.new
-
-    assert_equal 5, enigma.generate_key_string.length
-  end
-
-  def test_it_can_generate_keys
-    enigma = Enigma.new
-    key = "02715"
-
-    assert_equal ({
-      a: 2,
-      b: 27,
-      c: 71,
-      d: 15
-    }), enigma.generate_keys(key)
-  end
-
-  def test_it_can_generate_todays_date
-    enigma = Enigma.new
-
-    assert_equal Time.now.strftime("%d%m%y"), enigma.generate_date
-  end
-
-  def test_it_can_generate_offsets
-    enigma = Enigma.new
-    date = "040895"
-
-    assert_equal ({
-      a: 1,
-      b: 0,
-      c: 2,
-      d: 5
-      }), enigma.generate_offsets(date)
-  end
-
   def test_it_can_generate_shifts
     enigma = Enigma.new
     key = "02715"
